requirements.txt
fastapi>=0.104.1
uvicorn>=0.31.1
websockets>=12.0
redis>=5.0.1
chromadb>=0.4.22
python-jose>=3.3.0
python-multipart>=0.0.6
pydantic>=2.5.0
asyncio-mqtt>=0.16.1
httpx>=0.25.2
mcp>=1.13.0
PyJWT>=2.8.0
bcrypt>=4.1.2
